% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{loadModel}
\alias{loadModel}
\title{Load a model}
\usage{
loadModel(filename)
}
\arguments{
\item{filename}{path to model}
}
\value{
a model object
}
\description{
\code{loadModel} loads a model into copasi and returns a reference to it.
}
