% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{openCopasi}
\alias{openCopasi}
\title{Open the given model in Copasi}
\usage{
openCopasi(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\description{
Open the given model in Copasi
}
