% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCopasi.R
\docType{package}
\name{rCopasi-package}
\alias{rCopasi}
\alias{rCopasi-package}
\title{rCopasi: Copasi R API.}
\description{
Run timecourses and stuff
}
\section{quantity accessors}{

Use dataframes to manipulate species and quantities and stuff
}

\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/jonasfoe/copasi-r-api}
  \item Report bugs at \url{http://github.com/jonasfoe/copasi-r-api/issues}
}

}
\author{
\strong{Maintainer}: Jonas Förster \email{jonasfoe@gmail.com}

}
