% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{runSteadyState}
\alias{runSteadyState}
\title{Run to stready state}
\usage{
runSteadyState(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
a metabolite data frame
}
\description{
\code{runSteadyState} calculates the steady state and returns the species concentrations as a data frame
}
