% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimecourse}
\alias{runTimecourse}
\title{Run a timecourse}
\usage{
runTimecourse(duration, steps = 1000,
  initialtime = datamodel$getModel()$getInitialTime(),
  datamodel = pkg_env$curr_dm)
}
\arguments{
\item{duration}{timecourse duration}

\item{datamodel}{a model object}
}
\value{
a data frame with a time column and species concentration columns
}
\description{
\code{runTimecourse} runs a timecourse and returns the species data as a data frame
}
